feat : create guards
Create an local auth guard and its strategy, which allows connections if
the user is registered.

Create and jwt auth guard and its strategy which allows which allows the
passage if the user has the authorization.
The authorization is given thanks to a token, created with jwt.

Create a custom decorator which retrieves User entity.

Finally, add methods on authservice to implement the strategy.