fix : remove guards on the GET routes
It was a non sense that the users can't read all the message from any user
or find one by its id.

Before, an user can only read all the messages or
find one only if he is the author.

It makes more sense to have the reading actions accessible to all.
Editing actions (create, update, remove) remain accessible only
if the user is the author of the message.